cloud-functions/rec-generate-candidates/requirements.txt
functions-framework==3.*
requests
numpy
sentence_transformers
pandas
google-cloud-storage
firebase-admin